--- spec/controllers/signups_controller_spec.rb
@@ -1,33 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe SignupsController, type: :controller do
-  let(:user) { FactoryBot.create(:user) }
-  let(:user_params) { FactoryBot.attributes_for(:user) }
-
-  describe '#create' do
-    context '有効な属性値の場合' do
-      example 'ユーザを作成できること' do
-        expect { post :create, params: { user: user_params } }.to change(User, :count).by(1)
-      end
-
-      example 'ステータスが302であること' do
-        post :create, params: { user: user_params }
-        expect(response.status).to eq 302
-      end
-
-      example 'トップ画面に遷移すること' do
-        post :create, params: { user: user_params }
-        expect(response).to redirect_to root_url
-      end
-    end
-
-    context '無効な属性値の場合' do
-      example 'ユーザを作成できないこと' do
-        user_params = FactoryBot.attributes_for(:user, :invalid)
-        expect do
-          post :create, params: { user: user_params }
-        end.to_not change(User, :count)
-      end
-    end
-  end
-end
